# backend/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinimetria",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/copilot/analyze": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Analyze",
        "description": "🎯 Main endpoint: Natural language → Statistical analysis.",
        "operationId": "analyze_api_v2_copilot_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/plan": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Create Plan",
        "description": "📝 Stage 1: Create Analysis Plan (No Execution).",
        "operationId": "create_plan_api_v2_copilot_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/execute": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Execute Plan",
        "description": "🚀 Stage 2: Execute Analysis Plan.",
        "operationId": "execute_plan_api_v2_copilot_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/execute_stream": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Execute Plan Stream",
        "description": "🚀 Stage 2: Execute Analysis Plan with STREAMING output.\nReturns Server-Sent Events (SSE) for real-time progress.",
        "operationId": "execute_plan_stream_api_v2_copilot_execute_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/refine": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Refine",
        "description": "🔄 Refine existing analysis.\n\nExample:\n```\nPOST /api/v2/copilot/refine\n{\n    \"session_id\": \"xxx\",\n    \"refinement\": \"Добавь ROC-анализ для CRP\"\n}\n```",
        "operationId": "refine_api_v2_copilot_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/session/{session_id}": {
      "get": {
        "tags": [
          "Copilot"
        ],
        "summary": "Get Session",
        "description": "📋 Get session state for debugging/inspection.",
        "operationId": "get_session_api_v2_copilot_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/copilot/report": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Download Report",
        "description": "📄 Generate DOCX report from session results.",
        "operationId": "download_report_api_v2_copilot_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "request": {
            "type": "string",
            "title": "Request"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "request"
        ],
        "title": "AnalyzeRequest",
        "description": "Request to start new analysis.",
        "example": {
          "dataset_id": "abc123",
          "model": "google/gemini-2.5-flash",
          "request": "Сравни группы по исходу, проверь динамику лабораторных V1→V2, найди предикторы летальности"
        }
      },
...
      "ExecuteRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "plan_override": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Override"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ExecuteRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefineRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "refinement": {
            "type": "string",
            "title": "Refinement"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "refinement"
        ],
        "title": "RefineRequest"
      },
      "ReportRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "include_code": {
            "type": "boolean",
            "title": "Include Code",
            "default": true
          },
          "include_interpretation": {
            "type": "boolean",
            "title": "Include Interpretation",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ReportRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}